Refactor home page components and remove unused code
