+++ Breeze.Web/src/components/homePage/HomePage.tsx
@@ -1,10 +1,14 @@
 import AddButtons from './AddButtons'
 import Goals from './Goals'
 import Header from './Header'
-import BudgetSection from './budget/BudgetSection'
+import BudgetSection from './budgetSection/BudgetSection'
 import Progress from './Progress'
 import './HomePage.css'
 
+/**
+ * This is the home page component that calls the components that make up the home page
+ *
+ */
 const HomePage = () => {
 	return (
 		<div className='page'>
