+++ Breeze.Web/src/models/goal.ts
@@ -1,0 +1,4 @@
+export type Goal = {
+    id: number,
+    goal: string,
+}
